[XCPU]: Move post bus logic to Xenon context.
